shadowreader/requirements.txt
requests-futures==0.9.7
pytz==2018.4
boto3==1.7.4
PyYAML==3.12
